refactor: MCP tools modularization into submodules
