+++ server/databus_server.py
@@ -10,4 +10,3 @@ mcp.mount(network.mcp, namespace="network")
 
 if __name__ == "__main__":
     mcp.run()
-
