+++ server/tests/test_databus_client.py
@@ -1,9 +1,11 @@
 import pytest
 from utils.databus_client import DatabusClient, get_client
 
+
 @pytest.fixture
 def client():
     return DatabusClient("http://localhost:8000/", timeout=10.0)
+
 
 @pytest.mark.asyncio
 async def test_get_api_vehicles(client):
@@ -15,6 +17,7 @@ async def test_get_api_vehicles(client):
     assert "company" in vehicle
     assert "label" in vehicle
     assert "license_plate" in vehicle
+
 
 @pytest.mark.asyncio
 async def test_get_feed_vehicle_positions(client):
@@ -32,12 +35,13 @@ async def test_get_feed_vehicle_positions(client):
         assert "latitude" in entity["vehicle"]["position"]
         assert "longitude" in entity["vehicle"]["position"]
 
+
 @pytest.mark.asyncio
 async def test_get_api_routes(client):
     result = await client.get_api("routes")
     assert isinstance(result, list)
     assert len(result) > 0
-    
+
     route = result[0]
     assert "route_id" in route
     assert "route_short_name" in route
@@ -45,8 +49,7 @@ async def test_get_api_routes(client):
     assert "route_type" in route
 
 
-
 def test_singleton():
     client1 = get_client()
     client2 = get_client()
-    assert client1 is client2  # 
+    assert client1 is client2  #
