+++ server/topics/network/__init__.py
@@ -1,0 +1,7 @@
+from fastmcp import FastMCP
+
+mcp = FastMCP("network")
+
+from .history import *     
+from .network_overview import *  
+from .realtime import *
