+++ server/topics/operations/__init__.py
@@ -1,0 +1,8 @@
+from fastmcp import FastMCP
+
+mcp = FastMCP("operations")
+
+from .assignments import *     
+from .companies import *  
+from .operators import *
+from .routes import *
