+++ server/topics/vehicles/__init__.py
@@ -1,0 +1,7 @@
+from fastmcp import FastMCP
+
+mcp = FastMCP("vehicles")
+
+from .information import *     
+from .realtime import *  
+from .assignment import *
